requirements.txt
numpy
pandas
scikit-learn
pyyaml
nltk
dvc
